+++ BALSAMIC/constants/rules.py
@@ -58,6 +58,7 @@ SNAKEMAKE_RULES: Dict[str, Dict[str, list]] = {
             "snakemake_rules/annotation/varcaller_sv_filter.rule",
             "snakemake_rules/annotation/vcf2cytosure_convert.rule",
             "snakemake_rules/annotation/final_vcf_reheader.rule",
+            "snakemake_rules/annotation/msi_scan_reference.rule",
         ],
     },
     "single_targeted": {