+++ tests/conftest.py
@@ -610,7 +610,7 @@ def swegen_sv_frequency_path(reference_variants_dir_path: str) -> str:
 
 @pytest.fixture(scope="session")
 def msi_pon_path(reference_variants_dir_path: str) -> str:
-    """Create path for Swegen SVs."""
+    """Create MSI PON path"""
     return Path(reference_variants_dir_path, "msi_pon.list").as_posix()