+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Added:
 ^^^^^^
 * MSIsensor-pro container https://github.com/Clinical-Genomics/BALSAMIC/pull/1444
 * MSI analysis to the tumor-normal workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/1454
-* MSI analysis to the tumor-only workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/
+* MSI analysis to the tumor-only workflow https://github.com/Clinical-Genomics/BALSAMIC/pull/1463
 * Sentieon install directory path to case config arguments https://github.com/Clinical-Genomics/BALSAMIC/pull/1461
 
 Changed: